feat(chat): 新增会话历史面板样式和功能优化
新增 useTokenStatisticAnimation 和 useChatSession 钩子函数
优化 SessionHistoryPanel 组件样式和交互
添加 ChatView.css 样式文件
